Add search and browse option to chrome_browser.py and chrome_api.
Each method can be used with the "browse" or the "google" keyword.
The "browse" keyword enables a user to browse a specific website whereas
the "google" keyword enables a user to go a google search.

+++ modules/local_program.py
@@ -31,7 +31,8 @@ class LocalProgram(Module):
 
     def exit_program(self):
         """Terminate self.process (launched program)."""
-        self.process.terminate()
+        if self.process:
+            self.process.terminate()
 
     def run(self, **kwargs):
         """Run module by set kwargs and verb switch statement.
